Abstracted shared billing components for use in internal ui
