+++ pages/establishment/licence-fees/overview/views/index.jsx
@@ -1,75 +1,54 @@
 import React from 'react';
 import { useSelector } from 'react-redux';
-import { Snippet } from '@asl/components';
 import Layout from '../../views';
-import { numberWithCommas } from '../../../../../lib/utils';
-
-function SummarySection({ rows, title, currency }) {
-  return (
-    <table className="govuk-table">
-      <thead>
-        <tr>
-          <th colSpan={2}><Snippet>{title}</Snippet></th>
-        </tr>
-      </thead>
-      <tbody>
-        {
-          rows.map((row, index) => (
-            <tr key={index}>
-              <td><Snippet>{row.title}</Snippet></td>
-              <td className="pull-right">{row.currency && '£'}{numberWithCommas(row.value)}</td>
-            </tr>
-          ))
-        }
-      </tbody>
-    </table>
-  );
-}
+import SummaryTable from './summary-table';
 
 export default function Overview() {
   const fees = useSelector(state => state.static.fees);
+  const sections = [
+    {
+      title: 'fees.overview.establishment.title',
+      rows: [
+        {
+          title: 'fees.overview.establishment.fee',
+          value: fees.establishment,
+          currency: true
+        }
+      ]
+    },
+    {
+      title: 'fees.overview.personal.title',
+      rows: [
+        {
+          title: 'fees.overview.personal.count',
+          value: fees.numPersonal
+        },
+        {
+          title: 'fees.overview.personal.fee',
+          value: fees.personalFee,
+          currency: true
+        },
+        {
+          title: 'fees.overview.personal.total',
+          value: fees.personal,
+          currency: true
+        }
+      ]
+    },
+    {
+      title: 'fees.overview.total.title',
+      rows: [
+        {
+          title: 'fees.overview.total.fee',
+          value: fees.total,
+          currency: true
+        }
+      ]
+    }
+  ];
   return (
     <Layout tab={0}>
-      <h2><Snippet>title</Snippet></h2>
-      <SummarySection
-        title="fees.overview.establishment.title"
-        rows={[
-          {
-            title: 'fees.overview.establishment.fee',
-            value: fees.establishment,
-            currency: true
-          }
-        ]}
-      />
-      <SummarySection
-        title="fees.overview.personal.title"
-        rows={[
-          {
-            title: 'fees.overview.personal.count',
-            value: fees.numPersonal
-          },
-          {
-            title: 'fees.overview.personal.fee',
-            value: fees.personalFee,
-            currency: true
-          },
-          {
-            title: 'fees.overview.personal.total',
-            value: fees.personal,
-            currency: true
-          }
-        ]}
-      />
-      <SummarySection
-        title="fees.overview.total.title"
-        rows={[
-          {
-            title: 'fees.overview.total.fee',
-            value: fees.total,
-            currency: true
-          }
-        ]}
-      />
+      <SummaryTable sections={sections} />
     </Layout>
   );
 }
