+++ pages/establishment/licence-fees/overview/views/summary-table.jsx
@@ -1,0 +1,37 @@
+import React, { Fragment } from 'react';
+import { Snippet } from '@asl/components';
+import { numberWithCommas } from '../../../../../lib/utils';
+
+function SummarySection({ rows, title, currency }) {
+  return (
+    <table className="govuk-table">
+      <thead>
+        <tr>
+          <th colSpan={2}><Snippet>{title}</Snippet></th>
+        </tr>
+      </thead>
+      <tbody>
+        {
+          rows.map((row, index) => (
+            <tr key={index}>
+              <td><Snippet>{row.title}</Snippet></td>
+              <td className="pull-right">{row.currency && '£'}{numberWithCommas(row.value)}</td>
+            </tr>
+          ))
+        }
+      </tbody>
+    </table>
+  );
+}
+
+export default function SummaryTable({ sections }) {
+  return (
+    <Fragment>
+      {
+        sections.map((section, index) => (
+          <SummarySection key={index} title={section.title} rows={section.rows} />
+        ))
+      }
+    </Fragment>
+  );
+}
